+++ lib/src/ui/failure_messages.dart
@@ -32,6 +32,7 @@ String describeFailure(AppLocalizations l10n, Object error) {
     FailureCode.workerNotRunning => l10n.failureWorkerNotRunning,
     FailureCode.workerFailed => l10n.failureWorkerFailed(detail),
     FailureCode.pipelineStopped => l10n.failurePipelineStopped,
+    FailureCode.captureFailed => l10n.failureCaptureFailed(detail),
     FailureCode.windowsOnly => l10n.failureWindowsOnly,
     FailureCode.explorerUnsupported => l10n.failureExplorerUnsupported,
     FailureCode.downloadRejected => l10n.failureDownloadRejected(detail),